Rename project from whorl to covey
- Renamed main package directory from whorl/ to covey/
- Updated project name and description in pyproject.toml
- Updated all import statements across the codebase
- Updated README with new project name and usage examples
- Updated documentation and ADR files
- Updated configuration files (.nvimrc, .claude.md)
- Updated benchmark tests and CI scripts
- All tests passing with new name
