+++ README.md
@@ -1,6 +1,6 @@
-# Whorl - where branches and flowers spring forth
+# Covey - finding groups of similar code
 
-`whorl` is a CLI tool that detects similar code using treesitter AST analysis with locality-sensitive hashing:
+`covey` is a CLI tool that detects similar code using treesitter AST analysis with locality-sensitive hashing:
 - Find **duplicate** code blocks meaningful to the language (classes/functions), not just lines.
 - near-duplicates
 - structurally similar text.
@@ -22,16 +22,16 @@ Key flags:
 
 ```bash
 # Find exact duplicates (default threshold=1.0)
-whorl detect /path/to/codebase
+covey detect /path/to/codebase
 
 # Find near-duplicates with 80% similarity threshold
-whorl detect --threshold 0.8 /path/to/codebase
+covey detect --threshold 0.8 /path/to/codebase
 
 # Show diffs between similar blocks and use loose ruleset
-whorl --ruleset loose detect --diff --min-lines 10 /path/to/codebase
+covey --ruleset loose detect --diff --min-lines 10 /path/to/codebase
 
 # Output results in SARIF format for CI tools
-whorl detect --format sarif -o results.sarif /path/to/codebase
+covey detect --format sarif -o results.sarif /path/to/codebase
 ```
 
 ### Other sub commands
@@ -42,7 +42,7 @@ List all rules in a ruleset, along with their descriptions. Use `--language` to 
 
 #### treesitter
 
-Display how whorl normalizes source code into tree-sitter tokens for similarity detection -- helpful for debugging why a certain section of a file might be similar to another. Shows the original source code side-by-side with the normalized token representation.
+Display how covey normalizes source code into tree-sitter tokens for similarity detection -- helpful for debugging why a certain section of a file might be similar to another. Shows the original source code side-by-side with the normalized token representation.
 
 ## Dev setup
 
@@ -53,7 +53,7 @@ make test
 
 ## Benchmarks
 
-whorl includes a testing framework for comparing duplication detection tools against real-world codebases.
+covey includes a testing framework for comparing duplication detection tools against real-world codebases.
 
 ```bash
 # Run all benchmark tests
