+++ covey/models/__init__.py
@@ -1,5 +1,5 @@
 """Domain models for whorl."""
 
-from whorl.models.ast import ParsedFile, ParseResult
+from covey.models.ast import ParsedFile, ParseResult
 
 __all__ = ["ParsedFile", "ParseResult"]
