+++ covey/pipeline/verification.py
@@ -2,10 +2,10 @@ import logging
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from whorl.models.shingle import ShingledRegion
+from covey.models.shingle import ShingledRegion
 
 if TYPE_CHECKING:
-    from whorl.models.similarity import Region, SimilarRegionGroup
+    from covey.models.similarity import Region, SimilarRegionGroup
 
 
 logger = logging.getLogger(__name__)
@@ -206,7 +206,7 @@ def verify_similar_groups(
         )
 
         # Import here to avoid circular dependency
-        from whorl.models.similarity import SimilarRegionGroup
+        from covey.models.similarity import SimilarRegionGroup
 
         verified_group = SimilarRegionGroup(
             regions=group.regions,
