+++ covey/terminal_detect.py
