+++ tests/pipeline/languages/test_python.py
@@ -5,9 +5,9 @@ from pathlib import Path
 import pytest
 
 from tests.conftest import parse_fixture
-from whorl.pipeline.languages.python import PythonConfig
-from whorl.pipeline.region_extraction import extract_all_regions
-from whorl.pipeline.rules.engine import RuleEngine, build_default_rules, build_loose_rules
+from covey.pipeline.languages.python import PythonConfig
+from covey.pipeline.region_extraction import extract_all_regions
+from covey.pipeline.rules.engine import RuleEngine, build_default_rules, build_loose_rules
 
 
 # Fixture path
@@ -35,10 +35,10 @@ def test_language_name():
 
 def test_import_removal_rules():
     """Test that all import types (including future imports) are removed."""
-    from whorl.pipeline.parse import parse_source_code
-    from whorl.pipeline.shingle import ASTShingler
-    from whorl.pipeline.region_extraction import ExtractedRegion
-    from whorl.models.similarity import Region
+    from covey.pipeline.parse import parse_source_code
+    from covey.pipeline.shingle import ASTShingler
+    from covey.pipeline.region_extraction import ExtractedRegion
+    from covey.models.similarity import Region
 
     # Code with future imports, regular imports, and from imports
     source = b"""from __future__ import annotations
@@ -84,10 +84,10 @@ def foo():
 
 def test_type_checking_block_removal():
     """Test that TYPE_CHECKING blocks are removed."""
-    from whorl.pipeline.parse import parse_source_code
-    from whorl.pipeline.shingle import ASTShingler
-    from whorl.pipeline.region_extraction import ExtractedRegion
-    from whorl.models.similarity import Region
+    from covey.pipeline.parse import parse_source_code
+    from covey.pipeline.shingle import ASTShingler
+    from covey.pipeline.region_extraction import ExtractedRegion
+    from covey.models.similarity import Region
 
     # Code with TYPE_CHECKING block
     source = b"""import typing as t
@@ -129,10 +129,10 @@ def foo():
 
 def test_typevar_removal():
     """Test that TypeVar declarations are removed."""
-    from whorl.pipeline.parse import parse_source_code
-    from whorl.pipeline.shingle import ASTShingler
-    from whorl.pipeline.region_extraction import ExtractedRegion
-    from whorl.models.similarity import Region
+    from covey.pipeline.parse import parse_source_code
+    from covey.pipeline.shingle import ASTShingler
+    from covey.pipeline.region_extraction import ExtractedRegion
+    from covey.models.similarity import Region
 
     # Code with TypeVar declarations
     source = b"""import typing as t
